requirements.txt
aiohttp==3.7.4.post0
albumentations==1.0.3
alembic==1.6.5
aniso8601==9.0.1
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
asn1crypto==1.4.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
configparser==5.0.2
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
Flask==2.0.1
Flask-Admin==1.5.8
Flask-Bcrypt==0.7.1
Flask-Email==1.4.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gcloud==0.18.3
google==3.0.0
google-api-core==2.0.0
google-auth==2.0.1
google-cloud==0.34.0
google-cloud-speech==2.7.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.39.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
image==1.5.33
imageio==2.9.0
imgaug==0.4.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
libcst==0.3.20
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mpmath==1.2.1
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.0
numpy==1.20.3
oauth2client==4.1.3
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pg8000==1.21.1
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
proto-plus==1.19.0
protobuf==3.18.0rc1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
scramp==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.7.1
simpy==4.0.1
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.17
sqlparse==0.4.1
sympy==1.8
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.8.8
torch==1.9.0
torchsummary==1.5.1
torchvision==0.10.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
xlrd==2.0.1
yarl==1.6.3
